+++ src/freeton_crawler_db/main.ml
@@ -1,0 +1,19 @@
+(**************************************************************************)
+(*                                                                        *)
+(*  Copyright (c) 2021 OCamlPro SAS                                       *)
+(*                                                                        *)
+(*  All rights reserved.                                                  *)
+(*  This file is distributed under the terms of the GNU Lesser General    *)
+(*  Public License version 2.1, with the special exception on linking     *)
+(*  described in the LICENSE.md file in the root directory.               *)
+(*                                                                        *)
+(*                                                                        *)
+(**************************************************************************)
+
+
+(* If you delete or rename this file, you should add
+   'src/freeton_crawler_db/main.ml' to the 'skip' field in "drom.toml" *)
+
+let init = Updated.init
+
+let main () = Printf.printf "Hello world!\n"