+++ src/freeton_crawler_db_versions/main.mli
@@ -1,0 +1,5 @@
+
+val database : unit -> string
+
+val upgrades : (int * (unit PGOCaml.t -> int -> unit)) list
+val downgrades : (int * string list) list