+++ src/freeton_wallet_lib/main.ml
@@ -45,6 +45,7 @@ let main () =
       CommandUtils.cmd;
       CommandWatch.cmd;
       CommandInspect.cmd;
+      CommandCrawler.cmd;
     ]
   in
   let common_args =