+++ src/freeton_wallet_lib/processManager.mli
@@ -1,0 +1,17 @@
+(**************************************************************************)
+(*                                                                        *)
+(*  Copyright (c) 2021 OCamlPro SAS                                       *)
+(*                                                                        *)
+(*  All rights reserved.                                                  *)
+(*  This file is distributed under the terms of the GNU Lesser General    *)
+(*  Public License version 2.1, with the special exception on linking     *)
+(*  described in the LICENSE.md file in the root directory.               *)
+(*                                                                        *)
+(*                                                                        *)
+(**************************************************************************)
+
+val start : string -> string array -> unit Lwt.t
+val stop : string -> unit Lwt.t
+val status : string -> unit Lwt.t
+
+val set_worker : string -> unit Lwt.t