+++ CHANGES.md
@@ -1,4 +1,10 @@
 
-## v0.1.0 ( 2021-05-13 )
+## v0.3.1
+
+* Add 'ft crawler ACCOUNT' to crawle transactions on ACCOUNT and store
+  corresponding events in a PostgresQL database. Use --start/status/stop
+  to create a crawler with a monitoring process.
+
+## v0.1.0 ( 2021-05-14 )
 
 * Initial commit